chore: reorganize backend folder structure — Phase 3 EE mirror (REST API)
Mirror the Phase 3 REST API restructure inside `ee/`, staying entirely under
the Enterprise license boundary (no file leaves `ee/`):

- `ee/app/livechat-enterprise/server/api/` → `ee/server/api/v1/omnichannel/`
  (endpoints + `lib/` + relocated index aggregator)
- `ee/app/api-enterprise/server/` → `ee/server/api/` (canned-responses
  endpoint → `v1/`, `lib/` → `api/lib/`, `middlewares/` → `v1/middlewares/`)
- fold the `api-enterprise` index aggregator into `ee/server/index.ts` and
  delete it; repoint `livechat-enterprise/server/index.ts` at the new
  omnichannel aggregator.

Files move as-is; only import paths change. Verified with `yarn lint --quiet`
and `tsc --noEmit`.
