--- apps/meteor/ee/app/api-enterprise/server/index.ts
@@ -1,1 +1,0 @@
-import './canned-responses';
