+++ apps/meteor/ee/server/api/lib/canned-responses.js
@@ -1,8 +1,8 @@
 import { CannedResponse } from '@rocket.chat/models';
 import { escapeRegExp } from '@rocket.chat/string-helpers';
 
-import { hasPermissionAsync } from '../../../../../app/authorization/server/functions/hasPermission';
-import { getDepartmentsWhichUserCanAccess } from '../../../livechat-enterprise/server/api/lib/departments';
+import { hasPermissionAsync } from '../../../../app/authorization/server/functions/hasPermission';
+import { getDepartmentsWhichUserCanAccess } from '../v1/omnichannel/lib/departments';
 
 export async function findAllCannedResponses({ userId }) {
 	// If the user is an admin or livechat manager, get his own responses and all responses from all departments
