+++ apps/meteor/ee/server/api/v1/omnichannel/agents.ts
@@ -4,13 +4,13 @@ import {
 	isLivechatAnalyticsAgentsAvailableForServiceHistoryProps,
 } from '@rocket.chat/rest-typings';
 
-import { API } from '../../../../../app/api/server';
-import { getPaginationItems } from '../../../../../app/api/server/helpers/getPaginationItems';
 import {
 	findAllAverageServiceTimeAsync,
 	findAllServiceTimeAsync,
 	findAvailableServiceTimeHistoryAsync,
 } from '../../../../../app/livechat/server/lib/analytics/agents';
+import { API } from '../../../../../server/api';
+import { getPaginationItems } from '../../../../../server/api/lib/getPaginationItems';
 
 API.v1.addRoute(
 	'livechat/analytics/agents/average-service-time',
