+++ apps/meteor/ee/server/api/v1/omnichannel/business-hours.ts
@@ -1,9 +1,9 @@
 import type { ILivechatBusinessHour } from '@rocket.chat/core-typings';
 import type { PaginatedRequest } from '@rocket.chat/rest-typings';
 
-import { API } from '../../../../../app/api/server';
-import { getPaginationItems } from '../../../../../app/api/server/helpers/getPaginationItems';
-import { findBusinessHours } from '../business-hour/lib/business-hour';
+import { API } from '../../../../../server/api';
+import { getPaginationItems } from '../../../../../server/api/lib/getPaginationItems';
+import { findBusinessHours } from '../../../../app/livechat-enterprise/server/business-hour/lib/business-hour';
 
 declare module '@rocket.chat/rest-typings' {
 	// eslint-disable-next-line @typescript-eslint/naming-convention
