+++ apps/meteor/ee/server/api/v1/omnichannel/departments.ts
@@ -9,8 +9,6 @@ import {
 	isLivechatAnalyticsDepartmentsPercentageAbandonedChatsProps,
 } from '@rocket.chat/rest-typings';
 
-import { API } from '../../../../../app/api/server';
-import { getPaginationItems } from '../../../../../app/api/server/helpers/getPaginationItems';
 import {
 	findAllRoomsAsync,
 	findAllAverageServiceTimeAsync,
@@ -21,6 +19,8 @@ import {
 	findPercentageOfAbandonedRoomsAsync,
 	findAllAverageOfChatDurationTimeAsync,
 } from '../../../../../app/livechat/server/lib/analytics/departments';
+import { API } from '../../../../../server/api';
+import { getPaginationItems } from '../../../../../server/api/lib/getPaginationItems';
 
 API.v1.addRoute(
 	'livechat/analytics/departments/amount-of-chats',
