+++ apps/meteor/ee/server/api/v1/omnichannel/lib/definition.ts
