+++ apps/meteor/ee/server/api/v1/omnichannel/lib/priorities.ts
@@ -6,7 +6,7 @@ import { escapeRegExp } from '@rocket.chat/string-helpers';
 import type { FindOptions } from 'mongodb';
 
 import { notifyOnLivechatInquiryChangedByRoom, notifyOnRoomChanged } from '../../../../../../app/lib/server/lib/notifyListener';
-import { logger } from '../../lib/logger';
+import { logger } from '../../../../../app/livechat-enterprise/server/lib/logger';
 
 type FindPriorityParams = {
 	text?: string;
