+++ apps/meteor/ee/server/api/v1/omnichannel/monitors.ts
@@ -11,10 +11,10 @@ import {
 import { Meteor } from 'meteor/meteor';
 
 import { findMonitors, findMonitorByUsername } from './lib/monitors';
-import { API } from '../../../../../app/api/server';
-import type { ExtractRoutesFromAPI } from '../../../../../app/api/server/ApiClass';
-import { getPaginationItems } from '../../../../../app/api/server/helpers/getPaginationItems';
-import { LivechatEnterprise } from '../lib/LivechatEnterprise';
+import { API } from '../../../../../server/api';
+import type { ExtractRoutesFromAPI } from '../../../../../server/api/ApiClass';
+import { getPaginationItems } from '../../../../../server/api/lib/getPaginationItems';
+import { LivechatEnterprise } from '../../../../app/livechat-enterprise/server/lib/LivechatEnterprise';
 
 API.v1.addRoute(
 	'livechat/monitors',
