+++ apps/meteor/ee/server/api/v1/omnichannel/priorities.ts
@@ -2,9 +2,9 @@ import { LivechatPriority } from '@rocket.chat/models';
 import { isGETLivechatPrioritiesParams, isPUTLivechatPriority } from '@rocket.chat/rest-typings';
 
 import { findPriority, updatePriority } from './lib/priorities';
-import { API } from '../../../../../app/api/server';
-import { getPaginationItems } from '../../../../../app/api/server/helpers/getPaginationItems';
 import { notifyOnLivechatPriorityChanged } from '../../../../../app/lib/server/lib/notifyListener';
+import { API } from '../../../../../server/api';
+import { getPaginationItems } from '../../../../../server/api/lib/getPaginationItems';
 
 API.v1.addRoute(
 	'livechat/priorities',
