+++ apps/meteor/ee/server/api/v1/omnichannel/triggers.ts
@@ -4,8 +4,8 @@ import { isLivechatTriggerWebhookCallParams } from '@rocket.chat/rest-typings';
 import { isLivechatTriggerWebhookTestParams } from '@rocket.chat/rest-typings/src/v1/omnichannel';
 
 import { callTriggerExternalService } from './lib/triggers';
-import { API } from '../../../../../app/api/server';
 import { settings } from '../../../../../app/settings/server';
+import { API } from '../../../../../server/api';
 
 API.v1.addRoute(
 	'livechat/triggers/external-service/test',
